Using routers
All endpoints are now specified by FastAPI's APIRouter class.
This means that we can create the same endpoints several times, which is
useful for the different base URL's required by the OPTiMaDe
specification, e.g., /optimade, /optimade/v0.10, ...
