+++ optimade/server/routers/__init__.py
@@ -1,0 +1,4 @@
+from .references import references_coll
+from .structures import structures_coll
+
+ENTRY_COLLECTIONS = {"references": references_coll, "structures": structures_coll}
