+++ optimade/server/routers/info.py
@@ -1,0 +1,83 @@
+from typing import Union
+
+from fastapi import APIRouter
+from starlette.exceptions import HTTPException as StarletteHTTPException
+from starlette.requests import Request
+
+from optimade.models import (
+    ErrorResponse,
+    InfoResponse,
+    EntryInfoResponse,
+    ReferenceResource,
+    StructureResource,
+)
+
+from .utils import meta_values, retrieve_queryable_properties
+
+
+router = APIRouter()
+
+ENTRY_INFO_SCHEMAS = {
+    "structures": StructureResource.schema,
+    "references": ReferenceResource.schema,
+}
+
+
+@router.get(
+    "/info",
+    response_model=Union[InfoResponse, ErrorResponse],
+    response_model_skip_defaults=False,
+    tags=["Info"],
+)
+def get_info(request: Request):
+    from optimade.models import BaseInfoResource, BaseInfoAttributes
+
+    return InfoResponse(
+        meta=meta_values(str(request.url), 1, 1, more_data_available=False),
+        data=BaseInfoResource(
+            attributes=BaseInfoAttributes(
+                api_version="v0.10",
+                available_api_versions=[
+                    {
+                        "url": "http://localhost:5000/optimade/v0.10.0/",
+                        "version": "0.10.0",
+                    }
+                ],
+                entry_types_by_format={"json": ["structures"]},
+                available_endpoints=["info", "structures"],
+            )
+        ),
+    )
+
+
+@router.get(
+    "/info/{entry}",
+    response_model=Union[EntryInfoResponse, ErrorResponse],
+    response_model_skip_defaults=True,
+    tags=["Info", "Structure", "Reference"],
+)
+def get_entry_info(request: Request, entry: str):
+    from optimade.models import EntryInfoResource
+
+    valid_entry_info_endpoints = {"references", "structures"}
+    if entry not in valid_entry_info_endpoints:
+        raise StarletteHTTPException(
+            status_code=404,
+            detail=f"Entry info not found for {entry}, valid entry info endpoints are: {valid_entry_info_endpoints}",
+        )
+
+    schema = ENTRY_INFO_SCHEMAS[entry]()
+    queryable_properties = {"id", "type", "attributes"}
+    properties = retrieve_queryable_properties(schema, queryable_properties)
+
+    output_fields_by_format = {"json": list(properties.keys())}
+
+    return EntryInfoResponse(
+        meta=meta_values(str(request.url), 1, 1, more_data_available=False),
+        data=EntryInfoResource(
+            formats=list(output_fields_by_format.keys()),
+            description=schema.get("description", "Entry Resources"),
+            properties=properties,
+            output_fields_by_format=output_fields_by_format,
+        ),
+    )
